Making module make-card
